src/main/java/domain/rules/RockPaperScissors.java
package domain.rules;

import domain.Move;
import domain.Result;
import domain.Rules;

public class RockPaperScissors implements Rules {

    private Result getResultOfMatch(boolean result) {
        if (result) {
            return Result.WIN;
        }
        return Result.LOSE;
    }

    @Override
    public Result decideMatch(Move handOne, Move handTwo) {
        if (handOne.equals(handTwo)) {
            return Result.DRAW;
        }

        switch (handOne) {
            case ROCK:
                return getResultOfMatch(handTwo.equals(Move.SCISSORS));
            case PAPER:
                return getResultOfMatch(handTwo.equals(Move.ROCK));
            case SCISSORS:
                return getResultOfMatch(handTwo.equals(Move.PAPER));
        }

        return Result.DRAW;
    }
}
